refactor sales module into components
